executor_manager/pyproject.toml
[project]
name = "executor-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "boto3>=1.42.28",
    "docker>=7.1.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pyrefly]
project-includes = [
    "app/**/*.py",
    "app/**/*.pyi",
]

search_path = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
